[Fix] 필기 폭을 Double로 저장

+++ semomun/StudyVCs/ChildVCs/FormOne/MultipleWith5AnswerVC.swift
@@ -247,7 +247,7 @@ extension MultipleWith5AnswerVC: UICollectionViewDelegateFlowLayout {
 extension MultipleWith5AnswerVC {
     func canvasViewDrawingDidChange(_ canvasView: PKCanvasView) {
         let data = self.canvasView.drawing.dataRepresentation()
-        self.viewModel?.updatePagePencilData(to: data, width: Int(self.canvasView.contentSize.width))
+        self.viewModel?.updatePagePencilData(to: data, width: Double(self.canvasView.contentSize.width))
     }
 }
